update
update

+++ src/main/java/rex/service/Mapper/RecentlyPlayedMapper.java
@@ -15,7 +15,7 @@ public interface RecentlyPlayedMapper {
     List<Song> getRecentlyPlayedSongs(String userAccount);
 
     //增加用户最近播放的歌曲
-    void addRecentlyPlayedSong(int songId,String userAccount,String songName,String singer,int time,String coverUrl,String audioUrl,String lrcUrl,String album);
+    void addRecentlyPlayedSong(int songId,String userAccount,String songName,String singer,long time,String coverUrl,String audioUrl,String lrcUrl,String album);
 
     //删除用户最近播放的指定歌曲
     void deleteRecentlyPlayerSong(String userAccount,int songId);
